Fixed minor problems with the learning algorithms of the ffnnalfa network. Set the default number precision to 32. Updated examples code.

+++ library/brain.js
@@ -4,7 +4,7 @@ import * as m from "mathjs";
 
 m.config({
     number: "BigNumber",
-    precision: 64
+    precision: 32
 });
 
 import {Cerebrum} from "./cerebrum";
